requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.4.2
typing-extensions==4.7.1
loguru==0.7.0
aiohttp==3.9.1

# LLM and NLP
langchain==0.0.267
langchain-community==0.0.13
langchain-openai==0.0.2
sentencepiece==0.1.99
transformers==4.33.2
torch==2.0.1
spacy==3.6.1

# Vector database and embeddings
chroma-hnswlib==0.7.3
chromdb==0.4.18
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3

# Data processing
pyarrow==13.0.0
fastparquet==2023.7.0

# UI
streamlit==1.26.0
streamlit-chat==0.1.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0